candidateRepository implementation

+++ ZhetistikApp.Api/Interfaces/IAchievementRepository.cs
@@ -4,11 +4,11 @@ namespace ZhetistikApp.Api.Interfaces
 {
     public interface IAchievementRepository
     {
-        public Task<Achievement> GetAchievementByIdAsync(long id);
+        public Task<Achievement> GetAchievementByIdAsync(int id);
         public Task<IEnumerable<Achievement>> GetAllAchievementsAsync();
         public Task<IEnumerable<Achievement>> GetAchievementsByPersonAsync(string firstName, string lastName);
-        public Task<long> CreateAchievementAsync(Achievement achievement);
-        public Task DeleteAchievementAsync(long id);
+        public Task<int> CreateAchievementAsync(Achievement achievement);
+        public Task DeleteAchievementAsync(int id);
         public Task UpdateAchievementAsync(Achievement achievement);
     }
 }
