+++ ZhetistikApp.Api/Interfaces/IPlacementRepository.cs
@@ -4,11 +4,11 @@ namespace ZhetistikApp.Api.Interfaces
 {
     public interface IPlacementRepository
     {
-        public Task<Placement> GetPlacementAsync(long placementId);
-        public Task<Placement> GetPlacementAsync(long countryId, long cityId);
+        public Task<Placement> GetPlacementAsync(int placementId);
+        public Task<Placement> GetPlacementAsync(int countryId, int cityId);
         public Task<IEnumerable<Placement>> GetPlacementsAsync();
-        public Task<long> CreatePlacementAsync(Placement placement);
-        public Task DeletePlacementAsync(long id);
+        public Task<int> CreatePlacementAsync(Placement placement);
+        public Task DeletePlacementAsync(int id);
         public Task UpdatePlacementAsync();
     }
 }
