+++ ZhetistikApp.Api/Models/City.cs
@@ -7,7 +7,7 @@ namespace ZhetistikApp.Api.Models
     public class City
     {
         [Key]
-        public long CityID { get; set; }
+        public int CityID { get; set; }
         public string CityName { get; set; }
         public string PostalCode { get; set; }
     }
