+++ ZhetistikApp.Api/Models/Course.cs
@@ -7,7 +7,7 @@ namespace ZhetistikApp.Api.Models
     public class Course
     {
         [Key]
-        public long CourseID { get; set; }
+        public int CourseID { get; set; }
         public string CourseName { get; set; }
         public string CourseDescription { get; set; }
         public string CourseLength { get; set; }
