+++ ZhetistikApp.Api/Models/Portfolio.cs
@@ -7,10 +7,10 @@ namespace ZhetistikApp.Api.Models
     public class Portfolio
     {
         [Key]
-        public long PortfolioID { get; set; }
-        public long CandidateID { get; set; }
-        public long PlacementID { get; set; }
-        public long AchievementID { get; set; }
+        public int PortfolioID { get; set; }
+        public int CandidateID { get; set; }
+        public int PlacementID { get; set; }
+        public int AchievementID { get; set; }
         public DateTime CreatedDate { get; set; }
         public bool IsPublished { get; set; }
         public Portfolio()
