+++ ZhetistikApp.Api/Models/School.cs
@@ -7,8 +7,8 @@ namespace ZhetistikApp.Api.Models
     public class School
     {
         [Key]
-        public long SchoolID { get; set; }
-        public long PlacementID { get; set; }
+        public int SchoolID { get; set; }
+        public int PlacementID { get; set; }
         public string SchoolName { get; set; }
         public DateTime FoundationYear { get; set; }
     }
